refactored all Join methods and TestDataFrameRowOperations, removed appendRowToDataframe method
re implemented the appendRowToDataframe method and changed the name into AppendRow, replacing original AppendRow to be more robust in handling columns that do not currently exist in dataframe that is passed in. This fixed many inconsistent failing tests. Adding on, in OuterJoin, logic had issues where it did not append the rows that did not match in the previous for loops